requirements.txt
fastapi==0.115.2
uvicorn==0.34.3
python-multipart==0.0.20
sqlalchemy==2.0.41
pymupdf==1.26.0
llama-index==0.11.23
langchain==0.3.25
google-generativeai==0.8.5
pydantic==2.11.5
python-dotenv==1.1.0